+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 - `genomepy clean` wont complain when there is nothing to clean
+- properly gzip the annotation.gtf if it was unzipped during sanitizing.
 
 ## [0.9.0] - 2020-09-01